+++ Changes.md
@@ -8,6 +8,7 @@ you spot any mistakes.
 
 * Add `POOL_CLOSED` code to "Pool is closed." error
 * Add `POOL_CONNLIMIT` code to "No connections available." error #1332
+* Bind underlying socket to same domain as connection #1243
 * Fix edge cases constructing long stack traces #1387
 * Fix Query stream to emit close after ending #1349 #1350
 * Fix type cast for BIGINT columns when number is negative #1376